Extract common Toggl code to lib/toggl
