+++ toggl-freshbooks-invoice.rb
@@ -1,12 +1,10 @@
 #!/usr/bin/env ruby-1.9.3-p125@toggl-invoice
 
-require 'rest_client'
-require "base64"
-require 'yaml'
 require 'json'
+require 'rest_client'
 require 'time'
-require "rexml/document"
-
+require 'yaml'
+require_relative 'lib/toggl'
 
 def main
   $config = YAML.load_file(File.join(File.dirname(__FILE__), 'config', 'config.yml'))
@@ -32,62 +30,8 @@ def main
     puts
     puts 'A draft of your invoice has been created.'
     puts 'Log in to your FreshBooks account to review and send it.'
-	puts getFreshbooksUrl
+    puts getFreshbooksUrl
   end
-end
-
-# see https://github.com/toggl/toggl_api_docs/blob/master/reports/summary.md
-# and https://github.com/toggl/toggl_api_docs/blob/master/reports.md
-def getTogglSummary(date_since, date_until, client_name = nil)
-  params = {
-    'user_agent' => $config['company']['email'],
-    'workspace_id' => $config['toggl']['workspace'],
-    'since' => date_since,
-    'until' => date_until
-  }
-  params['client_ids'] = getTogglClientId(client_name) unless client_name.nil?
-
-  response = RestClient::Request.new(
-    :method => :get,
-    :url => 'https://toggl.com/reports/api/v2/summary',
-    :headers => {
-      :accept => :json,
-      :content_type => :json,
-      :authorization => 'Basic ' + Base64.urlsafe_encode64($config['toggl']['api_token'] + ':api_token'),
-      :params => params
-    }
-  ).execute
-
-  if response.code != 200
-    puts response.to_str
-    raise 'Error retrieving Toggl summary'
-  end
-  return response
-end
-
-# see https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-clients
-def getTogglClientId(client_name)
-  response = RestClient::Request.new(
-    :method => :get,
-    :url => 'https://www.toggl.com/api/v8/workspaces/' + $config['toggl']['workspace'] + '/clients',
-    :headers => { 
-      :accept => :json,
-      :content_type => :json,
-      :authorization => 'Basic ' + Base64.urlsafe_encode64($config['toggl']['api_token'] + ':api_token')
-    }
-  ).execute
-
-  if response.code != 200
-    puts response.to_str
-    raise 'Error retrieving Toggl clients'
-  end
-
-  JSON.parse(response.to_str).each do |client|
-    if client['name'] == client_name
-      return client['id']
-    end
-  end
-  raise 'Error retrieving Toggl client id for ' + client_name
 end
 
 def generateInvoiceData(toggl_summary, start_date, end_date, client_name)
